requirements.txt
aiogram==2.10.1
PyYAML==5.3.1
